Extract notify
